Added some initialization to backend
Added initialization shell script that runs on backend docker startup, inside is a variable USERNAME that sets the user name to be used. It must also be set in docker-compose so that the adversary is logged in as the correct user. I would imagine this to be set by the custom python backend program dynamically, along with the password.

Co-authored-by: Johan Valfridsson <gusvalfjo@student.gu.se>