Backend container folder structure
Each container for the backend now has its own shared folder with the host. In this folder we can put initialization scripts for the container. It also creates home directories with the name of the user of the ssh login. TODO is to clean up the code with Python Template Strings.

Co-Authored-By: Johan Valfridsson <gusvalfjo@student.gu.se>