Added acquire and release port
In main (for now) there one can acquire and release a port for a specified container, I think this is thread safe (as backend api might be async). Change range of ports according to backend capabilities. Also added feature so that one can get the port that is bound to a specific container, get_container_port().